Use common constant for ANE and LXA. Both instructions share the same circuit to produce the result (accumulator looped back to itself while using the special bus), so it can't be that these two constants are permanently different.
Sure the value varies across processors, temperature and other factors, but not
between these two instructions and not for a long time.

As for why 6510dtv core is using 0xee and 6510 core 0xff is beyond me.
Technically both are correct, in practice 0xee is more common but breaks that
one loader mentioned in the comment. I leave this inconsistency in for now.